libtraceevent: Remove TEP_PLUGIN_OPTIONS
When loading a plugin, the current API searches for TEP_PLUGIN_OPTIONS
array in plugin's symbols and loads any options from it. This API
duplicates:

 tep_plugin_add_options()

In order to simplify the APIs, The duplicated TEP_PLUGIN_OPTIONS is
removed as an API, as it is not used by any existing plugin.  The
tep_plugin_add_options() API is the only way to register plugin options.

Link: https://lore.kernel.org/20200122150002.763233-3-tz.stoyanov@gmail.com
Signed-off-by: Tzvetomir Stoyanov (VMware) <tz.stoyanov@gmail.com>
[ Also remove TEP_PLUGIN_ALIAS ]
Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>