libtraceevent: Handle case of unknown operator
The m68k architecture had a trace event with "m68k_memory[0].addr >> 13"
in it. The parsing after the ']' and before the '.' checked the priority
of the '.' but because it does not handle '.' as an operator it returned a
negative number. But process_op() did not check for an error and continued
processing. This caused a segfault.

Return immediately if the operator is not handled by get_op_prio().

Link: https://lore.kernel.org/all/20241119112850.219834f5@gandalf.local.home/

Link: https://lore.kernel.org/20241119151809.4da5f874@gandalf.local.home
Fixes: 6582b0aea1cc ("tools/events: Add files to create libtraceevent.a")
Reported-by: Jean-Michel Hautbois <jeanmichel.hautbois@yoseli.org>
Tested-by: Jean-Michel Hautbois <jeanmichel.hautbois@yoseli.org>
Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>