libtraceevent: Fix file endianness for big endian hosts
This patch initialises the file endianness to the same value as host endianness.
Without this, the file endianness is set to little endian by default, which
results in segfaults of libtracefs testsuite on big endian architectures
(libtracefs uses libtracevent, and my investigating led to this patch).
We can make a default assumption that the host and FS endianness is same.
If it is different, the user must set the correct endianness using the
event-parse-api (tep_set_file_bigendian)

After applying my fix, the run summary of utest/trace-utest of libtracefs:

Run Summary: Type Total Ran Passed Failed Inactive
              suites 1 1 n/a 0 0
               tests 36 36 35 1 0
             asserts 16407066 16407066 16407064 2 n/a

Elapsed time = 22.623 seconds

Link: https://lore.kernel.org/20241206054739.1431381-1-pragyansh.chaturvedi@canonical.com
Signed-off-by: Pragyansh Chaturvedi <pragyansh.chaturvedi@canonical.com>
Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>