libtraceevent: Return error on event parsing if format parsing fails
The event parsing should not fail if the "print fmt" portion fails, as
that is just a hint by the kernel on how to display the event. But that
portion can contain kernel internal information that is not available to
user space and there's no way the library can parse it properly. When that
fails, the event printing falls back to simply showing the content of all
the fields.

But if the fields themselves fail to parse, then there's no way to parse
the event properly. This should end up as a failure of parsing the event.

Link: https://lore.kernel.org/all/20250102174317.1594-1-shiju.jose@huawei.com/

Fixes: 6582b0aea1cc ("tools/events: Add files to create libtraceevent.a")
Link: https://lore.kernel.org/20250109102338.6128644d@gandalf.local.home
Reported-by: Shiju Jose <shiju.jose@huawei.com>
Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>